Create a dust odin model

Compile an odin model to work with dust.

odin_dust(x, ..., options = NULL)

odin_dust_(x, ..., options = NULL)

Arguments

x
Either the name of a file to read, a text string (if length is greater than 1 elements will be joined with newlines) or an expression.

...
Arguments passed to [odin::odin_dust_options], including `real_t`, `gpu`, `verbose`, `workdir`, `no_check_unused_equations` and `rewrite_dims`.

options
An [odin::odin_options] or `odin_dust_options` object. If given it overrides arguments; if it is already a `odin_dust_options` object it is returned unmodified. Otherwise it is passed through to [odin::odin_options] where it will override arguments in `...` but respond to the `odin_dust` specific options (`real_t`, etc)

Details

Note that this does not (yet) support the full odin output object, instead creating the more limited dust interface. However, for many uses this should be considerably faster than the interface that odin normally uses (built on dde).